Pull engagements.language.presetInventory from API when toggling preset inventory to keep cache fresh
We use a specific mutation operation to do this.
Note: Apollo currently thinks engagements are "loading" while this operation is in progress. I think this is a bug, we'll see if library updates fix it.
