+++ src/scenes/Projects/Overview/PresetInventory/TogglePresetInventory.graphql
@@ -1,0 +1,37 @@
+mutation TogglePresetInventory($input: UpdateProjectInput!) {
+  updateProject(input: $input) {
+    project {
+      ...RecalculatePresetInventory
+    }
+  }
+}
+
+fragment RecalculatePresetInventory on Project {
+  ...Id
+
+  presetInventory {
+    canEdit
+    canRead
+    value
+  }
+
+  engagements(input: { count: 100 }) {
+    # Skip pagination for now
+    items {
+      ...Id
+      ... on LanguageEngagement {
+        language {
+          value {
+            ...Id
+            presetInventory {
+              canRead
+              value
+            }
+          }
+        }
+      }
+    }
+  }
+
+  ...RecalculateChangesetDiff
+}
